+++ .env.sample
@@ -8,7 +8,8 @@ AZURE_OPENAI_TEMPERATURE=0
 AZURE_OPENAI_TOP_P=1.0
 AZURE_OPENAI_MAX_TOKENS=1000
 AZURE_OPENAI_STOP_SEQUENCE=
-AZURE_OPENAI_SYSTEM_MESSAGE=You are an AI assistant that helps people find information.
+# AZURE_OPENAI_SYSTEM_MESSAGE=You are an AI assistant that helps people find information.
+SYSTEM_MESSAGE=You are an AI assistant that helps people find information.
 AZURE_OPENAI_PREVIEW_API_VERSION=2024-02-15-preview
 AZURE_OPENAI_STREAM=True
 AZURE_OPENAI_ENDPOINT=
