+++ app.py
@@ -1,43 +1,43 @@
-import copy
-import json
+# import copy
+# import json
 import os
 import logging
-import uuid
+# import uuid
 from dotenv import load_dotenv
 from backend.auth.auth_enable import requires_auth
-import httpx
+# import httpx
 from quart import (
     Blueprint,
     Quart,
     jsonify,
-    make_response,
+    # make_response,
     request,
-    send_from_directory,
-    render_template,
+    # send_from_directory,
+    # render_template,
 )
-from openai import AsyncAzureOpenAI
-from azure.identity.aio import (
-    DefaultAzureCredential,
-    get_bearer_token_provider
-    )
-from backend.auth.auth_utils import get_authenticated_user_details
+# from openai import AsyncAzureOpenAI
+# from azure.identity.aio import (
+#     DefaultAzureCredential,
+#     get_bearer_token_provider
+#     )
+# from backend.auth.auth_utils import get_authenticated_user_details
 
-from backend.history.cosmosdbservice import CosmosConversationClient
+# from backend.history.cosmosdbservice import CosmosConversationClient
 
-from backend.utils import (
-    format_as_ndjson,
-    format_stream_response,
-    generateFilterString,
-    parse_multi_columns,
-    format_non_streaming_response,
-    convert_to_pf_format,
-    format_pf_non_streaming_response,
+from backend.azure_data.azure_file import (
+    conversation_internal
 )
+# from backend.azure_data.constants import CLOUD_PROVIDER
+
+from backend.aws_data.chatbot_api import KendraSearcher, kendra_searcher
+# from backend.aws_data.sm_chat import conversation_internal_sm
 
 bp = Blueprint("routes", __name__)
 
 # Current minimum Azure OpenAI version supported
-MINIMUM_SUPPORTED_AZURE_OPENAI_PREVIEW_API_VERSION = "2024-02-15-preview"
+# MINIMUM_SUPPORTED_AZURE_OPENAI_PREVIEW_API_VERSION = "2024-02-15-preview"
+
+CLOUD_PROVIDER=os.environ.get("CLOUD_PROVIDER")
 
 load_dotenv()
 
@@ -81,63 +81,63 @@ DEBUG = os.environ.get("DEBUG", "false")
 if DEBUG.lower() == "true":
     logging.basicConfig(level=logging.DEBUG)
 
-USER_AGENT = "GitHubSampleWebApp/AsyncAzureOpenAI/1.0.0"
+# USER_AGENT = "GitHubSampleWebApp/AsyncAzureOpenAI/1.0.0"
 
-# On Your Data Settings
-DATASOURCE_TYPE = os.environ.get("DATASOURCE_TYPE", "AzureCognitiveSearch")
-SEARCH_TOP_K = os.environ.get("SEARCH_TOP_K", 5)
-SEARCH_STRICTNESS = os.environ.get("SEARCH_STRICTNESS", 3)
-SEARCH_ENABLE_IN_DOMAIN = os.environ.get("SEARCH_ENABLE_IN_DOMAIN", "true")
+# # On Your Data Settings
+# DATASOURCE_TYPE = os.environ.get("DATASOURCE_TYPE", "AzureCognitiveSearch")
+# SEARCH_TOP_K = os.environ.get("SEARCH_TOP_K", 5)
+# SEARCH_STRICTNESS = os.environ.get("SEARCH_STRICTNESS", 3)
+# SEARCH_ENABLE_IN_DOMAIN = os.environ.get("SEARCH_ENABLE_IN_DOMAIN", "true")
 
-# ACS Integration Settings
-AZURE_SEARCH_SERVICE = os.environ.get("AZURE_SEARCH_SERVICE")
-AZURE_SEARCH_INDEX = os.environ.get("AZURE_SEARCH_INDEX")
-AZURE_SEARCH_KEY = os.environ.get("AZURE_SEARCH_KEY", None)
-AZURE_SEARCH_USE_SEMANTIC_SEARCH = os.environ.get(
-    "AZURE_SEARCH_USE_SEMANTIC_SEARCH", "false"
-)
-AZURE_SEARCH_SEMANTIC_SEARCH_CONFIG = os.environ.get(
-    "AZURE_SEARCH_SEMANTIC_SEARCH_CONFIG", "default"
-)
-AZURE_SEARCH_TOP_K = os.environ.get("AZURE_SEARCH_TOP_K", SEARCH_TOP_K)
-AZURE_SEARCH_ENABLE_IN_DOMAIN = os.environ.get(
-    "AZURE_SEARCH_ENABLE_IN_DOMAIN", SEARCH_ENABLE_IN_DOMAIN
-)
-AZURE_SEARCH_CONTENT_COLUMNS = os.environ.get("AZURE_SEARCH_CONTENT_COLUMNS")
-AZURE_SEARCH_FILENAME_COLUMN = os.environ.get("AZURE_SEARCH_FILENAME_COLUMN")
-AZURE_SEARCH_TITLE_COLUMN = os.environ.get("AZURE_SEARCH_TITLE_COLUMN")
-AZURE_SEARCH_URL_COLUMN = os.environ.get("AZURE_SEARCH_URL_COLUMN")
-AZURE_SEARCH_VECTOR_COLUMNS = os.environ.get("AZURE_SEARCH_VECTOR_COLUMNS")
-AZURE_SEARCH_QUERY_TYPE = os.environ.get("AZURE_SEARCH_QUERY_TYPE")
-AZURE_SEARCH_PERMITTED_GROUPS_COLUMN = os.environ.get(
-    "AZURE_SEARCH_PERMITTED_GROUPS_COLUMN"
-)
-AZURE_SEARCH_STRICTNESS = os.environ.get("AZURE_SEARCH_STRICTNESS", SEARCH_STRICTNESS)
+# # ACS Integration Settings
+# AZURE_SEARCH_SERVICE = os.environ.get("AZURE_SEARCH_SERVICE")
+# AZURE_SEARCH_INDEX = os.environ.get("AZURE_SEARCH_INDEX")
+# AZURE_SEARCH_KEY = os.environ.get("AZURE_SEARCH_KEY", None)
+# AZURE_SEARCH_USE_SEMANTIC_SEARCH = os.environ.get(
+#     "AZURE_SEARCH_USE_SEMANTIC_SEARCH", "false"
+# )
+# AZURE_SEARCH_SEMANTIC_SEARCH_CONFIG = os.environ.get(
+#     "AZURE_SEARCH_SEMANTIC_SEARCH_CONFIG", "default"
+# )
+# AZURE_SEARCH_TOP_K = os.environ.get("AZURE_SEARCH_TOP_K", SEARCH_TOP_K)
+# AZURE_SEARCH_ENABLE_IN_DOMAIN = os.environ.get(
+#     "AZURE_SEARCH_ENABLE_IN_DOMAIN", SEARCH_ENABLE_IN_DOMAIN
+# )
+# AZURE_SEARCH_CONTENT_COLUMNS = os.environ.get("AZURE_SEARCH_CONTENT_COLUMNS")
+# AZURE_SEARCH_FILENAME_COLUMN = os.environ.get("AZURE_SEARCH_FILENAME_COLUMN")
+# AZURE_SEARCH_TITLE_COLUMN = os.environ.get("AZURE_SEARCH_TITLE_COLUMN")
+# AZURE_SEARCH_URL_COLUMN = os.environ.get("AZURE_SEARCH_URL_COLUMN")
+# AZURE_SEARCH_VECTOR_COLUMNS = os.environ.get("AZURE_SEARCH_VECTOR_COLUMNS")
+# AZURE_SEARCH_QUERY_TYPE = os.environ.get("AZURE_SEARCH_QUERY_TYPE")
+# AZURE_SEARCH_PERMITTED_GROUPS_COLUMN = os.environ.get(
+#     "AZURE_SEARCH_PERMITTED_GROUPS_COLUMN"
+# )
+# AZURE_SEARCH_STRICTNESS = os.environ.get("AZURE_SEARCH_STRICTNESS", SEARCH_STRICTNESS)
 
-# AOAI Integration Settings
-AZURE_OPENAI_RESOURCE = os.environ.get("AZURE_OPENAI_RESOURCE")
-AZURE_OPENAI_MODEL = os.environ.get("AZURE_OPENAI_MODEL")
-AZURE_OPENAI_ENDPOINT = os.environ.get("AZURE_OPENAI_ENDPOINT")
-AZURE_OPENAI_KEY = os.environ.get("AZURE_OPENAI_KEY")
-AZURE_OPENAI_TEMPERATURE = os.environ.get("AZURE_OPENAI_TEMPERATURE", 0)
-AZURE_OPENAI_TOP_P = os.environ.get("AZURE_OPENAI_TOP_P", 1.0)
-AZURE_OPENAI_MAX_TOKENS = os.environ.get("AZURE_OPENAI_MAX_TOKENS", 1000)
-AZURE_OPENAI_STOP_SEQUENCE = os.environ.get("AZURE_OPENAI_STOP_SEQUENCE")
-AZURE_OPENAI_SYSTEM_MESSAGE = os.environ.get(
-    "AZURE_OPENAI_SYSTEM_MESSAGE",
-    "You are an AI assistant that helps people find information.",
-)
-AZURE_OPENAI_PREVIEW_API_VERSION = os.environ.get(
-    "AZURE_OPENAI_PREVIEW_API_VERSION",
-    MINIMUM_SUPPORTED_AZURE_OPENAI_PREVIEW_API_VERSION,
-)
-AZURE_OPENAI_STREAM = os.environ.get("AZURE_OPENAI_STREAM", "true")
-AZURE_OPENAI_MODEL_NAME = os.environ.get(
-    "AZURE_OPENAI_MODEL_NAME", "gpt-35-turbo-16k"
-)  # Name of the model, e.g. 'gpt-35-turbo-16k' or 'gpt-4'
-AZURE_OPENAI_EMBEDDING_ENDPOINT = os.environ.get("AZURE_OPENAI_EMBEDDING_ENDPOINT")
-AZURE_OPENAI_EMBEDDING_KEY = os.environ.get("AZURE_OPENAI_EMBEDDING_KEY")
-AZURE_OPENAI_EMBEDDING_NAME = os.environ.get("AZURE_OPENAI_EMBEDDING_NAME", "")
+# # AOAI Integration Settings
+# AZURE_OPENAI_RESOURCE = os.environ.get("AZURE_OPENAI_RESOURCE")
+# AZURE_OPENAI_MODEL = os.environ.get("AZURE_OPENAI_MODEL")
+# AZURE_OPENAI_ENDPOINT = os.environ.get("AZURE_OPENAI_ENDPOINT")
+# AZURE_OPENAI_KEY = os.environ.get("AZURE_OPENAI_KEY")
+# AZURE_OPENAI_TEMPERATURE = os.environ.get("AZURE_OPENAI_TEMPERATURE", 0)
+# AZURE_OPENAI_TOP_P = os.environ.get("AZURE_OPENAI_TOP_P", 1.0)
+# AZURE_OPENAI_MAX_TOKENS = os.environ.get("AZURE_OPENAI_MAX_TOKENS", 1000)
+# AZURE_OPENAI_STOP_SEQUENCE = os.environ.get("AZURE_OPENAI_STOP_SEQUENCE")
+# AZURE_OPENAI_SYSTEM_MESSAGE = os.environ.get(
+#     "AZURE_OPENAI_SYSTEM_MESSAGE",
+#     "You are an AI assistant that helps people find information.",
+# )
+# AZURE_OPENAI_PREVIEW_API_VERSION = os.environ.get(
+#     "AZURE_OPENAI_PREVIEW_API_VERSION",
+#     MINIMUM_SUPPORTED_AZURE_OPENAI_PREVIEW_API_VERSION,
+# )
+# AZURE_OPENAI_STREAM = os.environ.get("AZURE_OPENAI_STREAM", "true")
+# AZURE_OPENAI_MODEL_NAME = os.environ.get(
+#     "AZURE_OPENAI_MODEL_NAME", "gpt-35-turbo-16k"
+# )  # Name of the model, e.g. 'gpt-35-turbo-16k' or 'gpt-4'
+# AZURE_OPENAI_EMBEDDING_ENDPOINT = os.environ.get("AZURE_OPENAI_EMBEDDING_ENDPOINT")
+# AZURE_OPENAI_EMBEDDING_KEY = os.environ.get("AZURE_OPENAI_EMBEDDING_KEY")
+# AZURE_OPENAI_EMBEDDING_NAME = os.environ.get("AZURE_OPENAI_EMBEDDING_NAME", "")
 
 # # CosmosDB Mongo vcore vector db Settings
 # AZURE_COSMOSDB_MONGO_VCORE_CONNECTION_STRING = os.environ.get(
@@ -175,18 +175,18 @@ AZURE_OPENAI_EMBEDDING_NAME = os.environ.get("AZURE_OPENAI_EMBEDDING_NAME", "")
 #     "AZURE_COSMOSDB_MONGO_VCORE_VECTOR_COLUMNS"
 # )
 
-SHOULD_STREAM = True if AZURE_OPENAI_STREAM.lower() == "true" else False
+# SHOULD_STREAM = True if AZURE_OPENAI_STREAM.lower() == "true" else False
 
-# Chat History CosmosDB Integration Settings
-AZURE_COSMOSDB_DATABASE = os.environ.get("AZURE_COSMOSDB_DATABASE")
-AZURE_COSMOSDB_ACCOUNT = os.environ.get("AZURE_COSMOSDB_ACCOUNT")
-AZURE_COSMOSDB_CONVERSATIONS_CONTAINER = os.environ.get(
-    "AZURE_COSMOSDB_CONVERSATIONS_CONTAINER"
-)
-AZURE_COSMOSDB_ACCOUNT_KEY = os.environ.get("AZURE_COSMOSDB_ACCOUNT_KEY")
-AZURE_COSMOSDB_ENABLE_FEEDBACK = (
-    os.environ.get("AZURE_COSMOSDB_ENABLE_FEEDBACK", "false").lower() == "true"
-)
+# # Chat History CosmosDB Integration Settings
+# AZURE_COSMOSDB_DATABASE = os.environ.get("AZURE_COSMOSDB_DATABASE")
+# AZURE_COSMOSDB_ACCOUNT = os.environ.get("AZURE_COSMOSDB_ACCOUNT")
+# AZURE_COSMOSDB_CONVERSATIONS_CONTAINER = os.environ.get(
+#     "AZURE_COSMOSDB_CONVERSATIONS_CONTAINER"
+# )
+# AZURE_COSMOSDB_ACCOUNT_KEY = os.environ.get("AZURE_COSMOSDB_ACCOUNT_KEY")
+# AZURE_COSMOSDB_ENABLE_FEEDBACK = (
+#     os.environ.get("AZURE_COSMOSDB_ENABLE_FEEDBACK", "false").lower() == "true"
+# )
 
 # # Elasticsearch Integration Settings
 # ELASTICSEARCH_ENDPOINT = os.environ.get("ELASTICSEARCH_ENDPOINT")
@@ -237,48 +237,48 @@ AZURE_COSMOSDB_ENABLE_FEEDBACK = (
 # AZURE_MLINDEX_URL_COLUMN = os.environ.get("AZURE_MLINDEX_URL_COLUMN")
 # AZURE_MLINDEX_VECTOR_COLUMNS = os.environ.get("AZURE_MLINDEX_VECTOR_COLUMNS")
 # AZURE_MLINDEX_QUERY_TYPE = os.environ.get("AZURE_MLINDEX_QUERY_TYPE")
-# Promptflow Integration Settings
-USE_PROMPTFLOW = os.environ.get("USE_PROMPTFLOW", "false").lower() == "true"
-PROMPTFLOW_ENDPOINT = os.environ.get("PROMPTFLOW_ENDPOINT")
-PROMPTFLOW_API_KEY = os.environ.get("PROMPTFLOW_API_KEY")
-PROMPTFLOW_RESPONSE_TIMEOUT = os.environ.get("PROMPTFLOW_RESPONSE_TIMEOUT", 30.0)
-# default request and response field names are input -> 'query' and output -> 'reply'
-PROMPTFLOW_REQUEST_FIELD_NAME = os.environ.get("PROMPTFLOW_REQUEST_FIELD_NAME", "query")
-PROMPTFLOW_RESPONSE_FIELD_NAME = os.environ.get(
-    "PROMPTFLOW_RESPONSE_FIELD_NAME", "reply"
-)
-PROMPTFLOW_CITATIONS_FIELD_NAME = os.environ.get(
-    "PROMPTFLOW_CITATIONS_FIELD_NAME", "documents"
-)
-# Frontend Settings via Environment Variables
-AUTH_ENABLED = os.environ.get("AUTH_ENABLED", "true").lower() == "true"
-CHAT_HISTORY_ENABLED = (
-    AZURE_COSMOSDB_ACCOUNT
-    and AZURE_COSMOSDB_DATABASE
-    and AZURE_COSMOSDB_CONVERSATIONS_CONTAINER
-)
-SANITIZE_ANSWER = os.environ.get("SANITIZE_ANSWER", "false").lower() == "true"
-frontend_settings = {
-    "auth_enabled": AUTH_ENABLED,
-    "feedback_enabled": AZURE_COSMOSDB_ENABLE_FEEDBACK and CHAT_HISTORY_ENABLED,
-    # "ui": {
-    #     "title": UI_TITLE,
-    #     "logo": UI_LOGO,
-    #     "chat_logo": UI_CHAT_LOGO or UI_LOGO,
-    #     "chat_title": UI_CHAT_TITLE,
-    #     "chat_description": UI_CHAT_DESCRIPTION,
-    #     "show_share_button": UI_SHOW_SHARE_BUTTON,
-    # },
-    "sanitize_answer": SANITIZE_ANSWER,
-}
+# # Promptflow Integration Settings
+# USE_PROMPTFLOW = os.environ.get("USE_PROMPTFLOW", "false").lower() == "true"
+# PROMPTFLOW_ENDPOINT = os.environ.get("PROMPTFLOW_ENDPOINT")
+# PROMPTFLOW_API_KEY = os.environ.get("PROMPTFLOW_API_KEY")
+# PROMPTFLOW_RESPONSE_TIMEOUT = os.environ.get("PROMPTFLOW_RESPONSE_TIMEOUT", 30.0)
+# # default request and response field names are input -> 'query' and output -> 'reply'
+# PROMPTFLOW_REQUEST_FIELD_NAME = os.environ.get("PROMPTFLOW_REQUEST_FIELD_NAME", "query")
+# PROMPTFLOW_RESPONSE_FIELD_NAME = os.environ.get(
+#     "PROMPTFLOW_RESPONSE_FIELD_NAME", "reply"
+# )
+# PROMPTFLOW_CITATIONS_FIELD_NAME = os.environ.get(
+#     "PROMPTFLOW_CITATIONS_FIELD_NAME", "documents"
+# )
+# # Frontend Settings via Environment Variables
+# AUTH_ENABLED = os.environ.get("AUTH_ENABLED", "true").lower() == "true"
+# CHAT_HISTORY_ENABLED = (
+#     AZURE_COSMOSDB_ACCOUNT
+#     and AZURE_COSMOSDB_DATABASE
+#     and AZURE_COSMOSDB_CONVERSATIONS_CONTAINER
+# )
+# SANITIZE_ANSWER = os.environ.get("SANITIZE_ANSWER", "false").lower() == "true"
+# frontend_settings = {
+#     "auth_enabled": AUTH_ENABLED,
+#     "feedback_enabled": AZURE_COSMOSDB_ENABLE_FEEDBACK and CHAT_HISTORY_ENABLED,
+#     # "ui": {
+#     #     "title": UI_TITLE,
+#     #     "logo": UI_LOGO,
+#     #     "chat_logo": UI_CHAT_LOGO or UI_LOGO,
+#     #     "chat_title": UI_CHAT_TITLE,
+#     #     "chat_description": UI_CHAT_DESCRIPTION,
+#     #     "show_share_button": UI_SHOW_SHARE_BUTTON,
+#     # },
+#     "sanitize_answer": SANITIZE_ANSWER,
+# }
 
 
-def should_use_data():
-    global DATASOURCE_TYPE
-    if AZURE_SEARCH_SERVICE and AZURE_SEARCH_INDEX:
-        DATASOURCE_TYPE = "AzureCognitiveSearch"
-        logging.debug("Using Azure Cognitive Search")
-        return True
+# def should_use_data():
+#     global DATASOURCE_TYPE
+#     if AZURE_SEARCH_SERVICE and AZURE_SEARCH_INDEX:
+#         DATASOURCE_TYPE = "AzureCognitiveSearch"
+#         logging.debug("Using Azure Cognitive Search")
+#         return True
 
     # if (
     #     AZURE_COSMOSDB_MONGO_VCORE_DATABASE
@@ -305,877 +305,920 @@ def should_use_data():
     #     logging.debug("Using Azure ML Index")
     #     return True
 
-    return False
+    # return False
 
 
-SHOULD_USE_DATA = should_use_data()
+# SHOULD_USE_DATA = should_use_data()
 
 
 # Initialize Azure OpenAI Client
-def init_openai_client(use_data=SHOULD_USE_DATA):
-    azure_openai_client = None
-    try:
-        # API version check
-        if (
-            AZURE_OPENAI_PREVIEW_API_VERSION
-            < MINIMUM_SUPPORTED_AZURE_OPENAI_PREVIEW_API_VERSION
-        ):
-            raise Exception(
-                f"The minimum supported Azure OpenAI preview API version is '{MINIMUM_SUPPORTED_AZURE_OPENAI_PREVIEW_API_VERSION}'"
-            )
+# def init_openai_client(use_data=SHOULD_USE_DATA):
+#     azure_openai_client = None
+#     try:
+#         # API version check
+#         if (
+#             AZURE_OPENAI_PREVIEW_API_VERSION
+#             < MINIMUM_SUPPORTED_AZURE_OPENAI_PREVIEW_API_VERSION
+#         ):
+#             raise Exception(
+#                 f"The minimum supported Azure OpenAI preview API version is '{MINIMUM_SUPPORTED_AZURE_OPENAI_PREVIEW_API_VERSION}'"
+#             )
 
-        # Endpoint
-        if not AZURE_OPENAI_ENDPOINT and not AZURE_OPENAI_RESOURCE:
-            raise Exception(
-                "AZURE_OPENAI_ENDPOINT or AZURE_OPENAI_RESOURCE is required"
-            )
+#         # Endpoint
+#         if not AZURE_OPENAI_ENDPOINT and not AZURE_OPENAI_RESOURCE:
+#             raise Exception(
+#                 "AZURE_OPENAI_ENDPOINT or AZURE_OPENAI_RESOURCE is required"
+#             )
 
-        endpoint = (
-            AZURE_OPENAI_ENDPOINT
-            if AZURE_OPENAI_ENDPOINT
-            else f"https://{AZURE_OPENAI_RESOURCE}.openai.azure.com/"
-        )
+#         endpoint = (
+#             AZURE_OPENAI_ENDPOINT
+#             if AZURE_OPENAI_ENDPOINT
+#             else f"https://{AZURE_OPENAI_RESOURCE}.openai.azure.com/"
+#         )
 
-        # Authentication
-        aoai_api_key = AZURE_OPENAI_KEY
-        ad_token_provider = None
-        if not aoai_api_key:
-            logging.debug("No AZURE_OPENAI_KEY found, using Azure AD auth")
-            ad_token_provider = get_bearer_token_provider(
-                DefaultAzureCredential(), "https://cognitiveservices.azure.com/.default"
-            )
+#         # Authentication
+#         aoai_api_key = AZURE_OPENAI_KEY
+#         ad_token_provider = None
+#         if not aoai_api_key:
+#             logging.debug("No AZURE_OPENAI_KEY found, using Azure AD auth")
+#             ad_token_provider = get_bearer_token_provider(
+#                 DefaultAzureCredential(), "https://cognitiveservices.azure.com/.default"
+#             )
 
-        # Deployment
-        deployment = AZURE_OPENAI_MODEL
-        if not deployment:
-            raise Exception("AZURE_OPENAI_MODEL is required")
+#         # Deployment
+#         deployment = AZURE_OPENAI_MODEL
+#         if not deployment:
+#             raise Exception("AZURE_OPENAI_MODEL is required")
 
-        # Default Headers
-        default_headers = {"x-ms-useragent": USER_AGENT}
+#         # Default Headers
+#         default_headers = {"x-ms-useragent": USER_AGENT}
 
-        azure_openai_client = AsyncAzureOpenAI(
-            api_version=AZURE_OPENAI_PREVIEW_API_VERSION,
-            api_key=aoai_api_key,
-            azure_ad_token_provider=ad_token_provider,
-            default_headers=default_headers,
-            azure_endpoint=endpoint,
-        )
+#         azure_openai_client = AsyncAzureOpenAI(
+#             api_version=AZURE_OPENAI_PREVIEW_API_VERSION,
+#             api_key=aoai_api_key,
+#             azure_ad_token_provider=ad_token_provider,
+#             default_headers=default_headers,
+#             azure_endpoint=endpoint,
+#         )
 
-        return azure_openai_client
-    except Exception as e:
-        logging.exception("Exception in Azure OpenAI initialization", e)
-        azure_openai_client = None
-        raise e
-
-
-def init_cosmosdb_client():
-    cosmos_conversation_client = None
-    if CHAT_HISTORY_ENABLED:
-        try:
-            cosmos_endpoint = (
-                f"https://{AZURE_COSMOSDB_ACCOUNT}.documents.azure.com:443/"
-            )
-
-            if not AZURE_COSMOSDB_ACCOUNT_KEY:
-                credential = DefaultAzureCredential()
-            else:
-                credential = AZURE_COSMOSDB_ACCOUNT_KEY
-
-            cosmos_conversation_client = CosmosConversationClient(
-                cosmosdb_endpoint=cosmos_endpoint,
-                credential=credential,
-                database_name=AZURE_COSMOSDB_DATABASE,
-                container_name=AZURE_COSMOSDB_CONVERSATIONS_CONTAINER,
-                enable_message_feedback=AZURE_COSMOSDB_ENABLE_FEEDBACK,
-            )
-        except Exception as e:
-            logging.exception("Exception in CosmosDB initialization", e)
-            cosmos_conversation_client = None
-            raise e
-    else:
-        logging.debug("CosmosDB not configured")
-
-    return cosmos_conversation_client
+#         return azure_openai_client
+#     except Exception as e:
+#         logging.exception("Exception in Azure OpenAI initialization", e)
+#         azure_openai_client = None
+#         raise e
 
 
-def get_configured_data_source(application_id, run_id):
-    data_source = {}
-    query_type = "simple"
-    if DATASOURCE_TYPE == "AzureCognitiveSearch":
-        # Set query type
-        if AZURE_SEARCH_QUERY_TYPE:
-            query_type = AZURE_SEARCH_QUERY_TYPE
-        elif (
-            AZURE_SEARCH_USE_SEMANTIC_SEARCH.lower() == "true"
-            and AZURE_SEARCH_SEMANTIC_SEARCH_CONFIG
-        ):
-            query_type = "semantic"
+# def init_cosmosdb_client():
+#     cosmos_conversation_client = None
+#     if CHAT_HISTORY_ENABLED:
+#         try:
+#             cosmos_endpoint = (
+#                 f"https://{AZURE_COSMOSDB_ACCOUNT}.documents.azure.com:443/"
+#             )
 
-        # Set filter
-        # search_filter = (
-        #     "ApplicationID eq " + application_id +
-        #     " and RunID eq " + run_id
-        # )
+#             if not AZURE_COSMOSDB_ACCOUNT_KEY:
+#                 credential = DefaultAzureCredential()
+#             else:
+#                 credential = AZURE_COSMOSDB_ACCOUNT_KEY
 
-        search_filter = f"ApplicationID eq {application_id} and RunID eq {run_id}"
-        filter = None
-        userToken = None
-        if AZURE_SEARCH_PERMITTED_GROUPS_COLUMN:
-            userToken = request.headers.get("X-MS-TOKEN-AAD-ACCESS-TOKEN", "")
-            logging.debug(f"USER TOKEN is {'present' if userToken else 'not present'}")
-            if not userToken:
-                raise Exception(
-                    "Document-level access control is enabled, but user access token could not be fetched."
-                )
+#             cosmos_conversation_client = CosmosConversationClient(
+#                 cosmosdb_endpoint=cosmos_endpoint,
+#                 credential=credential,
+#                 database_name=AZURE_COSMOSDB_DATABASE,
+#                 container_name=AZURE_COSMOSDB_CONVERSATIONS_CONTAINER,
+#                 enable_message_feedback=AZURE_COSMOSDB_ENABLE_FEEDBACK,
+#             )
+#         except Exception as e:
+#             logging.exception("Exception in CosmosDB initialization", e)
+#             cosmos_conversation_client = None
+#             raise e
+#     else:
+#         logging.debug("CosmosDB not configured")
 
-            filter = generateFilterString(userToken)
-            # filter = "ApplicationID eq 116 and RunID eq 113"
-            # filter = "ApplicationID eq 108 and RunID eq 73"
-            logging.debug(f"FILTER: {filter}")
-            logging.debug(f"SEARCH FILTER: {search_filter}")
-
-        # Set authentication
-        authentication = {}
-        if AZURE_SEARCH_KEY:
-            authentication = {"type": "api_key", "api_key": AZURE_SEARCH_KEY}
-        else:
-            # If key is not provided, assume AOAI resource identity has been granted access to the search service
-            authentication = {"type": "system_assigned_managed_identity"}
-
-        data_source = {
-            "type": "azure_search",
-            "parameters": {
-                "endpoint": f"https://{AZURE_SEARCH_SERVICE}.search.windows.net",
-                "authentication": authentication,
-                "index_name": AZURE_SEARCH_INDEX,
-                "fields_mapping": {
-                    "content_fields": (
-                        parse_multi_columns(AZURE_SEARCH_CONTENT_COLUMNS)
-                        if AZURE_SEARCH_CONTENT_COLUMNS
-                        else []
-                    ),
-                    "title_field": (
-                        AZURE_SEARCH_TITLE_COLUMN if AZURE_SEARCH_TITLE_COLUMN else None
-                    ),
-                    "url_field": (
-                        AZURE_SEARCH_URL_COLUMN if AZURE_SEARCH_URL_COLUMN else None
-                    ),
-                    "filepath_field": (
-                        AZURE_SEARCH_FILENAME_COLUMN
-                        if AZURE_SEARCH_FILENAME_COLUMN
-                        else None
-                    ),
-                    "vector_fields": (
-                        parse_multi_columns(AZURE_SEARCH_VECTOR_COLUMNS)
-                        if AZURE_SEARCH_VECTOR_COLUMNS
-                        else []
-                    ),
-                },
-                "in_scope": (
-                    True if AZURE_SEARCH_ENABLE_IN_DOMAIN.lower() == "true" else False
-                ),
-                "top_n_documents": (
-                    int(AZURE_SEARCH_TOP_K) if AZURE_SEARCH_TOP_K else int(SEARCH_TOP_K)
-                ),
-                "query_type": query_type,
-                "semantic_configuration": (
-                    AZURE_SEARCH_SEMANTIC_SEARCH_CONFIG
-                    if AZURE_SEARCH_SEMANTIC_SEARCH_CONFIG
-                    else ""
-                ),
-                "role_information": AZURE_OPENAI_SYSTEM_MESSAGE,
-                "filter": search_filter,
-                "strictness": (
-                    int(AZURE_SEARCH_STRICTNESS)
-                    if AZURE_SEARCH_STRICTNESS
-                    else int(SEARCH_STRICTNESS)
-                ),
-            },
-        }
-
-    else:
-        raise Exception(
-            f"DATASOURCE_TYPE is not configured or unknown: {DATASOURCE_TYPE}"
-        )
-
-    if "vector" in query_type.lower() and DATASOURCE_TYPE != "AzureMLIndex":
-        embeddingDependency = {}
-        if AZURE_OPENAI_EMBEDDING_NAME:
-            embeddingDependency = {
-                "type": "deployment_name",
-                "deployment_name": AZURE_OPENAI_EMBEDDING_NAME,
-            }
-        elif AZURE_OPENAI_EMBEDDING_ENDPOINT and AZURE_OPENAI_EMBEDDING_KEY:
-            embeddingDependency = {
-                "type": "endpoint",
-                "endpoint": AZURE_OPENAI_EMBEDDING_ENDPOINT,
-                "authentication": {
-                    "type": "api_key",
-                    "key": AZURE_OPENAI_EMBEDDING_KEY,
-                },
-            }
-
-        else:
-            raise Exception(
-                f"Vector query type ({query_type}) is selected for data source type {DATASOURCE_TYPE} but no embedding dependency is configured"
-            )
-        data_source["parameters"]["embedding_dependency"] = embeddingDependency
-
-    return data_source
+#     return cosmos_conversation_client
 
 
-def prepare_model_args(request_body,application_id, run_id):
-    request_messages = request_body.get("messages", [])
-    messages = []
-    if not SHOULD_USE_DATA:
-        messages = [{"role": "system", "content": AZURE_OPENAI_SYSTEM_MESSAGE}]
+# def get_configured_data_source(application_id, run_id):
+#     data_source = {}
+#     query_type = AZURE_SEARCH_QUERY_TYPE  #"simple"
+#     if DATASOURCE_TYPE == "AzureCognitiveSearch":
+#         # Set query type
+#         if AZURE_SEARCH_QUERY_TYPE:
+#             query_type = AZURE_SEARCH_QUERY_TYPE
+#         elif (
+#             AZURE_SEARCH_USE_SEMANTIC_SEARCH.lower() == "true"
+#             and AZURE_SEARCH_SEMANTIC_SEARCH_CONFIG
+#         ):
+#             query_type = "semantic"
 
-    for message in request_messages:
-        if message:
-            messages.append({"role": message["role"], "content": message["content"]})
+#         # Set filter
+#         # search_filter = (
+#         #     "ApplicationID eq " + application_id +
+#         #     " and RunID eq " + run_id
+#         # )
 
-    model_args = {
-        "messages": messages,
-        "temperature": float(AZURE_OPENAI_TEMPERATURE),
-        "max_tokens": int(AZURE_OPENAI_MAX_TOKENS),
-        "top_p": float(AZURE_OPENAI_TOP_P),
-        "stop": (
-            parse_multi_columns(AZURE_OPENAI_STOP_SEQUENCE)
-            if AZURE_OPENAI_STOP_SEQUENCE
-            else None
-        ),
-        "stream": SHOULD_STREAM,
-        "model": AZURE_OPENAI_MODEL,
-    }
+#         search_filter = f"ApplicationID eq {application_id} and RunID eq {run_id}"
+#         filter = None
+#         userToken = None
+#         if AZURE_SEARCH_PERMITTED_GROUPS_COLUMN:
+#             userToken = request.headers.get("X-MS-TOKEN-AAD-ACCESS-TOKEN", "")
+#             logging.debug(f"USER TOKEN is {'present' if userToken else 'not present'}")
+#             if not userToken:
+#                 raise Exception(
+#                     "Document-level access control is enabled, but user access token could not be fetched."
+#                 )
 
-    if SHOULD_USE_DATA:
-        model_args["extra_body"] = {"data_sources": [get_configured_data_source(application_id, run_id)]}
+#             filter = generateFilterString(userToken)
+#             # filter = "ApplicationID eq 116 and RunID eq 113"
+#             # filter = "ApplicationID eq 108 and RunID eq 73"
+#             logging.debug(f"FILTER: {filter}")
+#             logging.debug(f"SEARCH FILTER: {search_filter}")
 
-    model_args_clean = copy.deepcopy(model_args)
-    if model_args_clean.get("extra_body"):
-        secret_params = [
-            "key",
-            "connection_string",
-            "embedding_key",
-            "encoded_api_key",
-            "api_key",
-        ]
-        for secret_param in secret_params:
-            if model_args_clean["extra_body"]["data_sources"][0]["parameters"].get(
-                secret_param
-            ):
-                model_args_clean["extra_body"]["data_sources"][0]["parameters"][
-                    secret_param
-                ] = "*****"
-        authentication = model_args_clean["extra_body"]["data_sources"][0][
-            "parameters"
-        ].get("authentication", {})
-        for field in authentication:
-            if field in secret_params:
-                model_args_clean["extra_body"]["data_sources"][0]["parameters"][
-                    "authentication"
-                ][field] = "*****"
-        embeddingDependency = model_args_clean["extra_body"]["data_sources"][0][
-            "parameters"
-        ].get("embedding_dependency", {})
-        if "authentication" in embeddingDependency:
-            for field in embeddingDependency["authentication"]:
-                if field in secret_params:
-                    model_args_clean["extra_body"]["data_sources"][0]["parameters"][
-                        "embedding_dependency"
-                    ]["authentication"][field] = "*****"
+#         # Set authentication
+#         authentication = {}
+#         if AZURE_SEARCH_KEY:
+#             authentication = {"type": "api_key", "api_key": AZURE_SEARCH_KEY}
+#         else:
+#             # If key is not provided, assume AOAI resource identity has been granted access to the search service
+#             authentication = {"type": "system_assigned_managed_identity"}
 
-    logging.debug(f"REQUEST BODY: {json.dumps(model_args_clean, indent=4)}")
+#         data_source = {
+#             "type": "azure_search",
+#             "parameters": {
+#                 "endpoint": f"https://{AZURE_SEARCH_SERVICE}.search.windows.net",
+#                 "authentication": authentication,
+#                 "index_name": AZURE_SEARCH_INDEX,
+#                 "fields_mapping": {
+#                     "content_fields": (
+#                         parse_multi_columns(AZURE_SEARCH_CONTENT_COLUMNS)
+#                         if AZURE_SEARCH_CONTENT_COLUMNS
+#                         else []
+#                     ),
+#                     "title_field": (
+#                         AZURE_SEARCH_TITLE_COLUMN if AZURE_SEARCH_TITLE_COLUMN else None
+#                     ),
+#                     "url_field": (
+#                         AZURE_SEARCH_URL_COLUMN if AZURE_SEARCH_URL_COLUMN else None
+#                     ),
+#                     "filepath_field": (
+#                         AZURE_SEARCH_FILENAME_COLUMN
+#                         if AZURE_SEARCH_FILENAME_COLUMN
+#                         else None
+#                     ),
+#                     "vector_fields": (
+#                         parse_multi_columns(AZURE_SEARCH_VECTOR_COLUMNS)
+#                         if AZURE_SEARCH_VECTOR_COLUMNS
+#                         else []
+#                     ),
+#                 },
+#                 "in_scope": (
+#                     True if AZURE_SEARCH_ENABLE_IN_DOMAIN.lower() == "true" else False
+#                 ),
+#                 "top_n_documents": (
+#                     int(AZURE_SEARCH_TOP_K) if AZURE_SEARCH_TOP_K else int(SEARCH_TOP_K)
+#                 ),
+#                 "query_type": query_type,
+#                 "semantic_configuration": (
+#                     AZURE_SEARCH_SEMANTIC_SEARCH_CONFIG
+#                     if AZURE_SEARCH_SEMANTIC_SEARCH_CONFIG
+#                     else ""
+#                 ),
+#                 "role_information": AZURE_OPENAI_SYSTEM_MESSAGE,
+#                 "filter": filter,
+#                 "strictness": (
+#                     int(AZURE_SEARCH_STRICTNESS)
+#                     if AZURE_SEARCH_STRICTNESS
+#                     else int(SEARCH_STRICTNESS)
+#                 ),
+#             },
+#         }
 
-    return model_args
+#     else:
+#         raise Exception(
+#             f"DATASOURCE_TYPE is not configured or unknown: {DATASOURCE_TYPE}"
+#         )
 
+#     if "vector" in query_type.lower() and DATASOURCE_TYPE != "AzureMLIndex":
+#         embeddingDependency = {}
+#         if AZURE_OPENAI_EMBEDDING_NAME:
+#             embeddingDependency = {
+#                 "type": "deployment_name",
+#                 "deployment_name": AZURE_OPENAI_EMBEDDING_NAME,
+#             }
+#         elif AZURE_OPENAI_EMBEDDING_ENDPOINT and AZURE_OPENAI_EMBEDDING_KEY:
+#             embeddingDependency = {
+#                 "type": "endpoint",
+#                 "endpoint": AZURE_OPENAI_EMBEDDING_ENDPOINT,
+#                 "authentication": {
+#                     "type": "api_key",
+#                     "key": AZURE_OPENAI_EMBEDDING_KEY,
+#                 },
+#             }
 
-async def promptflow_request(request,application_id, run_id):
-    try:
-        headers = {
-            "Content-Type": "application/json",
-            "Authorization": f"Bearer {PROMPTFLOW_API_KEY}",
-        }
-        # Adding timeout for scenarios where response takes longer to come back
-        logging.debug(f"Setting timeout to {PROMPTFLOW_RESPONSE_TIMEOUT}")
-        async with httpx.AsyncClient(
-            timeout=float(PROMPTFLOW_RESPONSE_TIMEOUT)
-        ) as client:
-            pf_formatted_obj = convert_to_pf_format(
-                request, application_id, run_id,
-                PROMPTFLOW_REQUEST_FIELD_NAME, 
-                PROMPTFLOW_RESPONSE_FIELD_NAME
-            )
-            # NOTE: This only support question and chat_history parameters
-            # If you need to add more parameters, you need to modify the request body
-            response = await client.post(
-                PROMPTFLOW_ENDPOINT,
-                json={
-                    f"{PROMPTFLOW_REQUEST_FIELD_NAME}": pf_formatted_obj[-1]["inputs"][
-                        PROMPTFLOW_REQUEST_FIELD_NAME
-                    ],
-                    "chat_history": pf_formatted_obj[:-1],
-                },
-                headers=headers,
-            )
-        resp = response.json()
-        resp["id"] = request["messages"][-1]["id"]
-        return resp
-    except Exception as e:
-        logging.error(f"An error occurred while making promptflow_request: {e}")
+#         else:
+#             raise Exception(
+#                 f"Vector query type ({query_type}) is selected for data source type {DATASOURCE_TYPE} but no embedding dependency is configured"
+#             )
+#         data_source["parameters"]["embedding_dependency"] = embeddingDependency
 
-
-async def send_chat_request(request, application_id, run_id):
-    filtered_messages = []
-    messages = request.get("messages", [])
-
-    for message in messages:
-        if message.get("role") != 'tool':
-            filtered_messages.append(message)
-
-    request['messages'] = filtered_messages
-    model_args = prepare_model_args(request,application_id, run_id)
-
-    try:
-        azure_openai_client = init_openai_client()
-        raw_response = await azure_openai_client.chat.completions.with_raw_response.create(**model_args)
-        response = raw_response.parse()
-        apim_request_id = raw_response.headers.get("apim-request-id")
-    except Exception as e:
-        logging.exception("Exception in send_chat_request")
-        raise e
-    
-    return response, apim_request_id
-
-
-async def complete_chat_request(request_body, application_id, run_id):
-    if USE_PROMPTFLOW and PROMPTFLOW_ENDPOINT and PROMPTFLOW_API_KEY:
-        response = await promptflow_request(request_body,application_id, run_id)
-        history_metadata = request_body.get("history_metadata", {})
-        return format_pf_non_streaming_response(
-            response, history_metadata,
-            PROMPTFLOW_RESPONSE_FIELD_NAME,
-            PROMPTFLOW_CITATIONS_FIELD_NAME
-        )
-    else:
-        response, apim_request_id = await send_chat_request(request_body,application_id, run_id)
-        history_metadata = request_body.get("history_metadata", {})
-        return format_non_streaming_response(
-            response,
-            history_metadata,
-            apim_request_id
-        )
+#     return data_source
 
 
-async def stream_chat_request(request_body, application_id, run_id):
-    response, apim_request_id = await send_chat_request(request_body, application_id, run_id)
-    history_metadata = request_body.get("history_metadata", {})
+# def prepare_model_args(request_body,application_id, run_id):
+#     request_messages = request_body.get("messages", [])
+#     messages = []
+#     if not SHOULD_USE_DATA:
+#         messages = [{"role": "system", "content": AZURE_OPENAI_SYSTEM_MESSAGE}]
 
-    async def generate():
-        async for completionChunk in response:
-            yield format_stream_response(
-                completionChunk,
-                history_metadata,
-                apim_request_id
-            )
+#     for message in request_messages:
+#         if message:
+#             messages.append({"role": message["role"], "content": message["content"]})
+#     model_args = {
+#         "messages": messages,
+#         "temperature": float(AZURE_OPENAI_TEMPERATURE),
+#         "max_tokens": int(AZURE_OPENAI_MAX_TOKENS),
+#         "top_p": float(AZURE_OPENAI_TOP_P),
+#         "stop": (
+#             parse_multi_columns(AZURE_OPENAI_STOP_SEQUENCE)
+#             if AZURE_OPENAI_STOP_SEQUENCE
+#             else None
+#         ),
+#         "stream": SHOULD_STREAM,
+#         "model": AZURE_OPENAI_MODEL,
+#     }
 
-    return generate()
+#     if SHOULD_USE_DATA:
+#         model_args["extra_body"] = {"data_sources": [get_configured_data_source(application_id, run_id)]}
+
+#     model_args_clean = copy.deepcopy(model_args)
+#     if model_args_clean.get("extra_body"):
+#         secret_params = [
+#             "key",
+#             "connection_string",
+#             "embedding_key",
+#             "encoded_api_key",
+#             "api_key",
+#         ]
+#         for secret_param in secret_params:
+#             if model_args_clean["extra_body"]["data_sources"][0]["parameters"].get(
+#                 secret_param
+#             ):
+#                 model_args_clean["extra_body"]["data_sources"][0]["parameters"][
+#                     secret_param
+#                 ] = "*****"
+#         authentication = model_args_clean["extra_body"]["data_sources"][0][
+#             "parameters"
+#         ].get("authentication", {})
+#         for field in authentication:
+#             if field in secret_params:
+#                 model_args_clean["extra_body"]["data_sources"][0]["parameters"][
+#                     "authentication"
+#                 ][field] = "*****"
+#         embeddingDependency = model_args_clean["extra_body"]["data_sources"][0][
+#             "parameters"
+#         ].get("embedding_dependency", {})
+#         if "authentication" in embeddingDependency:
+#             for field in embeddingDependency["authentication"]:
+#                 if field in secret_params:
+#                     model_args_clean["extra_body"]["data_sources"][0]["parameters"][
+#                         "embedding_dependency"
+#                     ]["authentication"][field] = "*****"
+
+#     logging.debug(f"REQUEST BODY: {json.dumps(model_args_clean, indent=4)}")
+
+#     return model_args
 
 
-async def conversation_internal(request_body, application_id, run_id):
-    try:
-        if SHOULD_STREAM:
-            result = await stream_chat_request(request_body, application_id, run_id)
-            response = await make_response(format_as_ndjson(result))
-            response.timeout = None
-            response.mimetype = "application/json-lines"
-            return response
-        else:
-            result = await complete_chat_request(request_body, application_id, run_id)
-            return jsonify(result)
+# async def promptflow_request(request,application_id, run_id):
+#     try:
+#         headers = {
+#             "Content-Type": "application/json",
+#             "Authorization": f"Bearer {PROMPTFLOW_API_KEY}",
+#         }
+#         # Adding timeout for scenarios where response takes longer to come back
+#         logging.debug(f"Setting timeout to {PROMPTFLOW_RESPONSE_TIMEOUT}")
+#         async with httpx.AsyncClient(
+#             timeout=float(PROMPTFLOW_RESPONSE_TIMEOUT)
+#         ) as client:
+#             pf_formatted_obj = convert_to_pf_format(
+#                 request, application_id, run_id,
+#                 PROMPTFLOW_REQUEST_FIELD_NAME, 
+#                 PROMPTFLOW_RESPONSE_FIELD_NAME
+#             )
+#             # NOTE: This only support question and chat_history parameters
+#             # If you need to add more parameters, you need to modify the request body
+#             response = await client.post(
+#                 PROMPTFLOW_ENDPOINT,
+#                 json={
+#                     f"{PROMPTFLOW_REQUEST_FIELD_NAME}": pf_formatted_obj[-1]["inputs"][
+#                         PROMPTFLOW_REQUEST_FIELD_NAME
+#                     ],
+#                     "chat_history": pf_formatted_obj[:-1],
+#                 },
+#                 headers=headers,
+#             )
+#         resp = response.json()
+#         resp["id"] = request["messages"][-1]["id"]
+#         return resp
+#     except Exception as e:
+#         logging.error(f"An error occurred while making promptflow_request: {e}")
 
-    except Exception as ex:
-        logging.exception(ex)
-        if hasattr(ex, "status_code"):
-            return jsonify({"error": str(ex)}), ex.status_code
-        else:
-            return jsonify({"error": str(ex)}), 500
+
+# async def send_chat_request(request, application_id, run_id):
+#     filtered_messages = []
+#     messages = request.get("messages", [])
+
+#     for message in messages:
+#         if message.get("role") != 'tool':
+#             filtered_messages.append(message)
+
+#     request['messages'] = filtered_messages
+#     model_args = prepare_model_args(request,application_id, run_id)
+
+#     try:
+#         azure_openai_client = init_openai_client()
+#         raw_response = await azure_openai_client.chat.completions.with_raw_response.create(**model_args)
+#         response = raw_response.parse()
+#         apim_request_id = raw_response.headers.get("apim-request-id")
+#     except Exception as e:
+#         logging.exception("Exception in send_chat_request")
+#         raise e
+#     return response, apim_request_id
+
+
+# async def complete_chat_request(request_body, application_id, run_id):
+#     if USE_PROMPTFLOW and PROMPTFLOW_ENDPOINT and PROMPTFLOW_API_KEY:
+#         response = await promptflow_request(request_body,application_id, run_id)
+#         history_metadata = request_body.get("history_metadata", {})
+#         return format_pf_non_streaming_response(
+#             response, history_metadata,
+#             PROMPTFLOW_RESPONSE_FIELD_NAME,
+#             PROMPTFLOW_CITATIONS_FIELD_NAME
+#         )
+#     else:
+#         response, apim_request_id = await send_chat_request(request_body,application_id, run_id)
+#         history_metadata = request_body.get("history_metadata", {})
+#         return format_non_streaming_response(
+#             response,
+#             history_metadata,
+#             apim_request_id
+#         )
+
+
+# async def stream_chat_request(request_body, application_id, run_id):
+#     response, apim_request_id = await send_chat_request(request_body, application_id, run_id)
+#     history_metadata = request_body.get("history_metadata", {})
+
+#     async def generate():
+#         async for completionChunk in response:
+#             yield format_stream_response(
+#                 completionChunk,
+#                 history_metadata,
+#                 apim_request_id
+#             )
+
+#     return generate()
+
+
+# async def conversation_internal(request_body, application_id, run_id):
+#     try:
+#         if SHOULD_STREAM:
+#             result = await stream_chat_request(request_body, application_id, run_id)
+#             response = await make_response(format_as_ndjson(result))
+#             response.timeout = None
+#             response.mimetype = "application/json-lines"
+#             return response
+#         else:
+#             result = await complete_chat_request(request_body, application_id, run_id)
+#             return jsonify(result)
+
+#     except Exception as ex:
+#         logging.exception(ex)
+#         if hasattr(ex, "status_code"):
+#             return jsonify({"error": str(ex)}), ex.status_code
+#         else:
+#             return jsonify({"error": str(ex)}), 500
+
+# ===========================================================================================
+# @bp.route("/conversation/<string:application_id>/<string:run_id>", methods=["POST"])
+# @requires_auth(roles=["userRole","admin"])
+# async def conversation(application_id:str,run_id:str):
+#     if CLOUD_PROVIDER=="azure":
+#         if not request.is_json:
+#             return jsonify({"error": "request must be json"}), 415
+#         request_json = await request.get_json()
+#         return await conversation_internal(request_json,application_id,run_id)
+#     else:
+#         pass
+# =======================================================================================================
 
 
 @bp.route("/conversation/<string:application_id>/<string:run_id>", methods=["POST"])
-# @requires_auth(roles=["userRole"])
+# @requires_auth(roles=["userRole","admin","adminRole"])
 async def conversation(application_id:str,run_id:str):
-    if not request.is_json:
-        return jsonify({"error": "request must be json"}), 415
-    request_json = await request.get_json()
-    return await conversation_internal(request_json,application_id,run_id)
+    if CLOUD_PROVIDER=="azure":
+        if not request.is_json:
+            return jsonify({"error": "request must be json"}), 415
+        request_json = await request.get_json()
+        print("request_json=",request_json)  # request_json= {'messages': [{'role': 'user', 'content': 'Hi'}]}
+        return await conversation_internal(request_json,application_id,run_id) 
+    elif CLOUD_PROVIDER=="aws":
+        try:
+            if application_id is None or run_id is None:
+                print("Missing required parameters")
+                return jsonify({"error": "Missing required parameters"},400)
+    
+            if not request.is_json:
+                return jsonify({"error": "request must be json"},415)
+            request_json = await request.get_json()
+            # print("request_json aws=",request_json)
+            # return await conversation_internal(request_json,application_id,run_id)
 
-## Conversation History API ##
-@bp.route("/history/generate", methods=["POST"])
-async def add_conversation():
-    authenticated_user = get_authenticated_user_details(request_headers=request.headers)
-    user_id = authenticated_user["user_principal_id"]
-
-    ## check request for conversation_id
-    request_json = await request.get_json()
-    conversation_id = request_json.get("conversation_id", None)
-
-    try:
-        # make sure cosmos is configured
-        cosmos_conversation_client = init_cosmosdb_client()
-        if not cosmos_conversation_client:
-            raise Exception("CosmosDB is not configured or not working")
-
-        # check for the conversation_id, if the conversation is not set, we will create a new one
-        history_metadata = {}
-        if not conversation_id:
-            title = await generate_title(request_json["messages"])
-            conversation_dict = await cosmos_conversation_client.create_conversation(
-                user_id=user_id, title=title
-            )
-            conversation_id = conversation_dict["id"]
-            history_metadata["title"] = title
-            history_metadata["date"] = conversation_dict["createdAt"]
-
-        ## Format the incoming message object in the "chat/completions" messages format
-        ## then write it to the conversation history in cosmos
-        messages = request_json["messages"]
-        if len(messages) > 0 and messages[-1]["role"] == "user":
-            createdMessageValue = await cosmos_conversation_client.create_message(
-                uuid=str(uuid.uuid4()),
-                conversation_id=conversation_id,
-                user_id=user_id,
-                input_message=messages[-1],
-            )
-            if createdMessageValue == "Conversation not found":
-                raise Exception(
-                    "Conversation not found for the given conversation ID: "
-                    + conversation_id
-                    + "."
-                )
+        except Exception as e:
+            print(f"Error parsing JSON: {e}")
+            return jsonify({"error": "Invalid JSON request"},400)
+        
+        response = await kendra_searcher.search(request_json, application_id, run_id)
+        # print("response AWS = ",response)
+        if (response is None):
+            logging.debug("Data not found in application")
+            return jsonify({"error": "Data not found in application"},404)
         else:
-            raise Exception("No user message found")
+            return response
 
-        await cosmos_conversation_client.cosmosdb_client.close()
-
-        # Submit request to Chat Completions for response
-        request_body = await request.get_json()
-        history_metadata["conversation_id"] = conversation_id
-        request_body["history_metadata"] = history_metadata
-        return await conversation_internal(request_body)
-
-    except Exception as e:
-        logging.exception("Exception in /history/generate")
-        return jsonify({"error": str(e)}), 500
+        # return await conversation_internal_sm(request_json,application_id,run_id) 
+# ====================================================================================================
 
 
-@bp.route("/history/update", methods=["POST"])
-async def update_conversation():
-    authenticated_user = get_authenticated_user_details(request_headers=request.headers)
-    user_id = authenticated_user["user_principal_id"]
-
-    ## check request for conversation_id
-    request_json = await request.get_json()
-    conversation_id = request_json.get("conversation_id", None)
-
-    try:
-        # make sure cosmos is configured
-        cosmos_conversation_client = init_cosmosdb_client()
-        if not cosmos_conversation_client:
-            raise Exception("CosmosDB is not configured or not working")
-
-        # check for the conversation_id, if the conversation is not set, we will create a new one
-        if not conversation_id:
-            raise Exception("No conversation_id found")
-
-        ## Format the incoming message object in the "chat/completions" messages format
-        ## then write it to the conversation history in cosmos
-        messages = request_json["messages"]
-        if len(messages) > 0 and messages[-1]["role"] == "assistant":
-            if len(messages) > 1 and messages[-2].get("role", None) == "tool":
-                # write the tool message first
-                await cosmos_conversation_client.create_message(
-                    uuid=str(uuid.uuid4()),
-                    conversation_id=conversation_id,
-                    user_id=user_id,
-                    input_message=messages[-2],
-                )
-            # write the assistant message
-            await cosmos_conversation_client.create_message(
-                uuid=messages[-1]["id"],
-                conversation_id=conversation_id,
-                user_id=user_id,
-                input_message=messages[-1],
-            )
-        else:
-            raise Exception("No bot messages found")
-
-        # Submit request to Chat Completions for response
-        await cosmos_conversation_client.cosmosdb_client.close()
-        response = {"success": True}
-        return jsonify(response), 200
-
-    except Exception as e:
-        logging.exception("Exception in /history/update")
-        return jsonify({"error": str(e)}), 500
 
 
-@bp.route("/history/message_feedback", methods=["POST"])
-async def update_message():
-    authenticated_user = get_authenticated_user_details(request_headers=request.headers)
-    user_id = authenticated_user["user_principal_id"]
-    cosmos_conversation_client = init_cosmosdb_client()
 
-    ## check request for message_id
-    request_json = await request.get_json()
-    message_id = request_json.get("message_id", None)
-    message_feedback = request_json.get("message_feedback", None)
-    try:
-        if not message_id:
-            return jsonify({"error": "message_id is required"}), 400
+# ## Conversation History API ##
+# @bp.route("/history/generate", methods=["POST"])
+# async def add_conversation():
+#     authenticated_user = get_authenticated_user_details(request_headers=request.headers)
+#     user_id = authenticated_user["user_principal_id"]
 
-        if not message_feedback:
-            return jsonify({"error": "message_feedback is required"}), 400
+#     ## check request for conversation_id
+#     request_json = await request.get_json()
+#     conversation_id = request_json.get("conversation_id", None)
 
-        ## update the message in cosmos
-        updated_message = await cosmos_conversation_client.update_message_feedback(
-            user_id, message_id, message_feedback
-        )
-        if updated_message:
-            return (
-                jsonify(
-                    {
-                        "message": f"Successfully updated message with feedback {message_feedback}",
-                        "message_id": message_id,
-                    }
-                ),
-                200,
-            )
-        else:
-            return (
-                jsonify(
-                    {
-                        "error": f"Unable to update message {message_id}. It either does not exist or the user does not have access to it."
-                    }
-                ),
-                404,
-            )
+#     try:
+#         # make sure cosmos is configured
+#         cosmos_conversation_client = init_cosmosdb_client()
+#         if not cosmos_conversation_client:
+#             raise Exception("CosmosDB is not configured or not working")
 
-    except Exception as e:
-        logging.exception("Exception in /history/message_feedback")
-        return jsonify({"error": str(e)}), 500
+#         # check for the conversation_id, if the conversation is not set, we will create a new one
+#         history_metadata = {}
+#         if not conversation_id:
+#             title = await generate_title(request_json["messages"])
+#             conversation_dict = await cosmos_conversation_client.create_conversation(
+#                 user_id=user_id, title=title
+#             )
+#             conversation_id = conversation_dict["id"]
+#             history_metadata["title"] = title
+#             history_metadata["date"] = conversation_dict["createdAt"]
 
+#         ## Format the incoming message object in the "chat/completions" messages format
+#         ## then write it to the conversation history in cosmos
+#         messages = request_json["messages"]
+#         if len(messages) > 0 and messages[-1]["role"] == "user":
+#             createdMessageValue = await cosmos_conversation_client.create_message(
+#                 uuid=str(uuid.uuid4()),
+#                 conversation_id=conversation_id,
+#                 user_id=user_id,
+#                 input_message=messages[-1],
+#             )
+#             if createdMessageValue == "Conversation not found":
+#                 raise Exception(
+#                     "Conversation not found for the given conversation ID: "
+#                     + conversation_id
+#                     + "."
+#                 )
+#         else:
+#             raise Exception("No user message found")
 
-@bp.route("/history/delete", methods=["DELETE"])
-async def delete_conversation():
-    ## get the user id from the request headers
-    authenticated_user = get_authenticated_user_details(request_headers=request.headers)
-    user_id = authenticated_user["user_principal_id"]
+#         await cosmos_conversation_client.cosmosdb_client.close()
 
-    ## check request for conversation_id
-    request_json = await request.get_json()
-    conversation_id = request_json.get("conversation_id", None)
+#         # Submit request to Chat Completions for response
+#         request_body = await request.get_json()
+#         history_metadata["conversation_id"] = conversation_id
+#         request_body["history_metadata"] = history_metadata
+#         return await conversation_internal(request_body)
 
-    try:
-        if not conversation_id:
-            return jsonify({"error": "conversation_id is required"}), 400
-
-        ## make sure cosmos is configured
-        cosmos_conversation_client = init_cosmosdb_client()
-        if not cosmos_conversation_client:
-            raise Exception("CosmosDB is not configured or not working")
-
-        ## delete the conversation messages from cosmos first
-        deleted_messages = await cosmos_conversation_client.delete_messages(
-            conversation_id, user_id
-        )
-
-        ## Now delete the conversation
-        deleted_conversation = await cosmos_conversation_client.delete_conversation(
-            user_id, conversation_id
-        )
-
-        await cosmos_conversation_client.cosmosdb_client.close()
-
-        return (
-            jsonify(
-                {
-                    "message": "Successfully deleted conversation and messages",
-                    "conversation_id": conversation_id,
-                }
-            ),
-            200,
-        )
-    except Exception as e:
-        logging.exception("Exception in /history/delete")
-        return jsonify({"error": str(e)}), 500
+#     except Exception as e:
+#         logging.exception("Exception in /history/generate")
+#         return jsonify({"error": str(e)}), 500
 
 
-@bp.route("/history/list", methods=["GET"])
-async def list_conversations():
-    offset = request.args.get("offset", 0)
-    authenticated_user = get_authenticated_user_details(request_headers=request.headers)
-    user_id = authenticated_user["user_principal_id"]
+# @bp.route("/history/update", methods=["POST"])
+# async def update_conversation():
+#     authenticated_user = get_authenticated_user_details(request_headers=request.headers)
+#     user_id = authenticated_user["user_principal_id"]
 
-    ## make sure cosmos is configured
-    cosmos_conversation_client = init_cosmosdb_client()
-    if not cosmos_conversation_client:
-        raise Exception("CosmosDB is not configured or not working")
+#     ## check request for conversation_id
+#     request_json = await request.get_json()
+#     conversation_id = request_json.get("conversation_id", None)
 
-    ## get the conversations from cosmos
-    conversations = await cosmos_conversation_client.get_conversations(
-        user_id, offset=offset, limit=25
-    )
-    await cosmos_conversation_client.cosmosdb_client.close()
-    if not isinstance(conversations, list):
-        return jsonify({"error": f"No conversations for {user_id} were found"}), 404
+#     try:
+#         # make sure cosmos is configured
+#         cosmos_conversation_client = init_cosmosdb_client()
+#         if not cosmos_conversation_client:
+#             raise Exception("CosmosDB is not configured or not working")
 
-    ## return the conversation ids
+#         # check for the conversation_id, if the conversation is not set, we will create a new one
+#         if not conversation_id:
+#             raise Exception("No conversation_id found")
 
-    return jsonify(conversations), 200
+#         ## Format the incoming message object in the "chat/completions" messages format
+#         ## then write it to the conversation history in cosmos
+#         messages = request_json["messages"]
+#         if len(messages) > 0 and messages[-1]["role"] == "assistant":
+#             if len(messages) > 1 and messages[-2].get("role", None) == "tool":
+#                 # write the tool message first
+#                 await cosmos_conversation_client.create_message(
+#                     uuid=str(uuid.uuid4()),
+#                     conversation_id=conversation_id,
+#                     user_id=user_id,
+#                     input_message=messages[-2],
+#                 )
+#             # write the assistant message
+#             await cosmos_conversation_client.create_message(
+#                 uuid=messages[-1]["id"],
+#                 conversation_id=conversation_id,
+#                 user_id=user_id,
+#                 input_message=messages[-1],
+#             )
+#         else:
+#             raise Exception("No bot messages found")
+
+#         # Submit request to Chat Completions for response
+#         await cosmos_conversation_client.cosmosdb_client.close()
+#         response = {"success": True}
+#         return jsonify(response), 200
+
+#     except Exception as e:
+#         logging.exception("Exception in /history/update")
+#         return jsonify({"error": str(e)}), 500
 
 
-@bp.route("/history/read", methods=["POST"])
-async def get_conversation():
-    authenticated_user = get_authenticated_user_details(request_headers=request.headers)
-    user_id = authenticated_user["user_principal_id"]
+# @bp.route("/history/message_feedback", methods=["POST"])
+# async def update_message():
+#     authenticated_user = get_authenticated_user_details(request_headers=request.headers)
+#     user_id = authenticated_user["user_principal_id"]
+#     cosmos_conversation_client = init_cosmosdb_client()
 
-    ## check request for conversation_id
-    request_json = await request.get_json()
-    conversation_id = request_json.get("conversation_id", None)
+#     ## check request for message_id
+#     request_json = await request.get_json()
+#     message_id = request_json.get("message_id", None)
+#     message_feedback = request_json.get("message_feedback", None)
+#     try:
+#         if not message_id:
+#             return jsonify({"error": "message_id is required"}), 400
 
-    if not conversation_id:
-        return jsonify({"error": "conversation_id is required"}), 400
+#         if not message_feedback:
+#             return jsonify({"error": "message_feedback is required"}), 400
 
-    ## make sure cosmos is configured
-    cosmos_conversation_client = init_cosmosdb_client()
-    if not cosmos_conversation_client:
-        raise Exception("CosmosDB is not configured or not working")
+#         ## update the message in cosmos
+#         updated_message = await cosmos_conversation_client.update_message_feedback(
+#             user_id, message_id, message_feedback
+#         )
+#         if updated_message:
+#             return (
+#                 jsonify(
+#                     {
+#                         "message": f"Successfully updated message with feedback {message_feedback}",
+#                         "message_id": message_id,
+#                     }
+#                 ),
+#                 200,
+#             )
+#         else:
+#             return (
+#                 jsonify(
+#                     {
+#                         "error": f"Unable to update message {message_id}. It either does not exist or the user does not have access to it."
+#                     }
+#                 ),
+#                 404,
+#             )
 
-    ## get the conversation object and the related messages from cosmos
-    conversation = await cosmos_conversation_client.get_conversation(
-        user_id, conversation_id
-    )
-    ## return the conversation id and the messages in the bot frontend format
-    if not conversation:
-        return (
-            jsonify(
-                {
-                    "error": f"Conversation {conversation_id} was not found. It either does not exist or the logged in user does not have access to it."
-                }
-            ),
-            404,
-        )
+#     except Exception as e:
+#         logging.exception("Exception in /history/message_feedback")
+#         return jsonify({"error": str(e)}), 500
+
+
+# @bp.route("/history/delete", methods=["DELETE"])
+# async def delete_conversation():
+#     ## get the user id from the request headers
+#     authenticated_user = get_authenticated_user_details(request_headers=request.headers)
+#     user_id = authenticated_user["user_principal_id"]
+
+#     ## check request for conversation_id
+#     request_json = await request.get_json()
+#     conversation_id = request_json.get("conversation_id", None)
+
+#     try:
+#         if not conversation_id:
+#             return jsonify({"error": "conversation_id is required"}), 400
+
+#         ## make sure cosmos is configured
+#         cosmos_conversation_client = init_cosmosdb_client()
+#         if not cosmos_conversation_client:
+#             raise Exception("CosmosDB is not configured or not working")
+
+#         ## delete the conversation messages from cosmos first
+#         deleted_messages = await cosmos_conversation_client.delete_messages(
+#             conversation_id, user_id
+#         )
+
+#         ## Now delete the conversation
+#         deleted_conversation = await cosmos_conversation_client.delete_conversation(
+    #         user_id, conversation_id
+    #     )
+
+    #     await cosmos_conversation_client.cosmosdb_client.close()
+
+    #     return (
+    #         jsonify(
+    #             {
+    #                 "message": "Successfully deleted conversation and messages",
+    #                 "conversation_id": conversation_id,
+    #             }
+    #         ),
+    #         200,
+    #     )
+    # except Exception as e:
+    #     logging.exception("Exception in /history/delete")
+    #     return jsonify({"error": str(e)}), 500
+
+
+# @bp.route("/history/list", methods=["GET"])
+# async def list_conversations():
+#     offset = request.args.get("offset", 0)
+#     authenticated_user = get_authenticated_user_details(request_headers=request.headers)
+#     user_id = authenticated_user["user_principal_id"]
+
+#     ## make sure cosmos is configured
+#     cosmos_conversation_client = init_cosmosdb_client()
+#     if not cosmos_conversation_client:
+#         raise Exception("CosmosDB is not configured or not working")
+
+#     ## get the conversations from cosmos
+#     conversations = await cosmos_conversation_client.get_conversations(
+#         user_id, offset=offset, limit=25
+#     )
+#     await cosmos_conversation_client.cosmosdb_client.close()
+#     if not isinstance(conversations, list):
+#         return jsonify({"error": f"No conversations for {user_id} were found"}), 404
+
+#     ## return the conversation ids
+
+#     return jsonify(conversations), 200
+
+
+# @bp.route("/history/read", methods=["POST"])
+# async def get_conversation():
+#     authenticated_user = get_authenticated_user_details(request_headers=request.headers)
+#     user_id = authenticated_user["user_principal_id"]
+
+#     ## check request for conversation_id
+#     request_json = await request.get_json()
+#     conversation_id = request_json.get("conversation_id", None)
+
+#     if not conversation_id:
+#         return jsonify({"error": "conversation_id is required"}), 400
+
+#     ## make sure cosmos is configured
+#     cosmos_conversation_client = init_cosmosdb_client()
+#     if not cosmos_conversation_client:
+#         raise Exception("CosmosDB is not configured or not working")
+
+#     ## get the conversation object and the related messages from cosmos
+#     conversation = await cosmos_conversation_client.get_conversation(
+#         user_id, conversation_id
+#     )
+#     ## return the conversation id and the messages in the bot frontend format
+#     if not conversation:
+#         return (
+#             jsonify(
+#                 {
+#                     "error": f"Conversation {conversation_id} was not found. It either does not exist or the logged in user does not have access to it."
+#                 }
+#             ),
+#             404,
+#         )
 
     # get the messages for the conversation from cosmos
-    conversation_messages = await cosmos_conversation_client.get_messages(
-        user_id, conversation_id
-    )
+    # conversation_messages = await cosmos_conversation_client.get_messages(
+    #     user_id, conversation_id
+    # )
 
-    ## format the messages in the bot frontend format
-    messages = [
-        {
-            "id": msg["id"],
-            "role": msg["role"],
-            "content": msg["content"],
-            "createdAt": msg["createdAt"],
-            "feedback": msg.get("feedback"),
-        }
-        for msg in conversation_messages
-    ]
+    # ## format the messages in the bot frontend format
+    # messages = [
+    #     {
+    #         "id": msg["id"],
+    #         "role": msg["role"],
+    #         "content": msg["content"],
+    #         "createdAt": msg["createdAt"],
+    #         "feedback": msg.get("feedback"),
+    #     }
+    #     for msg in conversation_messages
+    # ]
 
-    await cosmos_conversation_client.cosmosdb_client.close()
-    return jsonify({"conversation_id": conversation_id, "messages": messages}), 200
-
-
-@bp.route("/history/rename", methods=["POST"])
-async def rename_conversation():
-    authenticated_user = get_authenticated_user_details(request_headers=request.headers)
-    user_id = authenticated_user["user_principal_id"]
-
-    ## check request for conversation_id
-    request_json = await request.get_json()
-    conversation_id = request_json.get("conversation_id", None)
-
-    if not conversation_id:
-        return jsonify({"error": "conversation_id is required"}), 400
-
-    ## make sure cosmos is configured
-    cosmos_conversation_client = init_cosmosdb_client()
-    if not cosmos_conversation_client:
-        raise Exception("CosmosDB is not configured or not working")
-
-    ## get the conversation from cosmos
-    conversation = await cosmos_conversation_client.get_conversation(
-        user_id, conversation_id
-    )
-    if not conversation:
-        return (
-            jsonify(
-                {
-                    "error": f"Conversation {conversation_id} was not found. It either does not exist or the logged in user does not have access to it."
-                }
-            ),
-            404,
-        )
-
-    ## update the title
-    title = request_json.get("title", None)
-    if not title:
-        return jsonify({"error": "title is required"}), 400
-    conversation["title"] = title
-    updated_conversation = await cosmos_conversation_client.upsert_conversation(
-        conversation
-    )
-
-    await cosmos_conversation_client.cosmosdb_client.close()
-    return jsonify(updated_conversation), 200
+    # await cosmos_conversation_client.cosmosdb_client.close()
+    # return jsonify({"conversation_id": conversation_id, "messages": messages}), 200
 
 
-@bp.route("/history/delete_all", methods=["DELETE"])
-async def delete_all_conversations():
-    ## get the user id from the request headers
-    authenticated_user = get_authenticated_user_details(request_headers=request.headers)
-    user_id = authenticated_user["user_principal_id"]
+# @bp.route("/history/rename", methods=["POST"])
+# async def rename_conversation():
+#     authenticated_user = get_authenticated_user_details(request_headers=request.headers)
+#     user_id = authenticated_user["user_principal_id"]
 
-    # get conversations for user
-    try:
-        ## make sure cosmos is configured
-        cosmos_conversation_client = init_cosmosdb_client()
-        if not cosmos_conversation_client:
-            raise Exception("CosmosDB is not configured or not working")
+#     ## check request for conversation_id
+#     request_json = await request.get_json()
+#     conversation_id = request_json.get("conversation_id", None)
 
-        conversations = await cosmos_conversation_client.get_conversations(
-            user_id, offset=0, limit=None
-        )
-        if not conversations:
-            return jsonify({"error": f"No conversations for {user_id} were found"}), 404
+#     if not conversation_id:
+#         return jsonify({"error": "conversation_id is required"}), 400
 
-        # delete each conversation
-        for conversation in conversations:
-            ## delete the conversation messages from cosmos first
-            deleted_messages = await cosmos_conversation_client.delete_messages(
-                conversation["id"], user_id
-            )
+#     ## make sure cosmos is configured
+#     cosmos_conversation_client = init_cosmosdb_client()
+#     if not cosmos_conversation_client:
+#         raise Exception("CosmosDB is not configured or not working")
 
-            ## Now delete the conversation
-            deleted_conversation = await cosmos_conversation_client.delete_conversation(
-                user_id, conversation["id"]
-            )
-        await cosmos_conversation_client.cosmosdb_client.close()
-        return (
-            jsonify(
-                {
-                    "message": f"Successfully deleted conversation and messages for user {user_id}"
-                }
-            ),
-            200,
-        )
+#     ## get the conversation from cosmos
+#     conversation = await cosmos_conversation_client.get_conversation(
+#         user_id, conversation_id
+    # )
+    # if not conversation:
+    #     return (
+    #         jsonify(
+    #             {
+    #                 "error": f"Conversation {conversation_id} was not found. It either does not exist or the logged in user does not have access to it."
+    #             }
+    #         ),
+    #         404,
+    #     )
 
-    except Exception as e:
-        logging.exception("Exception in /history/delete_all")
-        return jsonify({"error": str(e)}), 500
+    # ## update the title
+    # title = request_json.get("title", None)
+    # if not title:
+    #     return jsonify({"error": "title is required"}), 400
+    # conversation["title"] = title
+    # updated_conversation = await cosmos_conversation_client.upsert_conversation(
+    #     conversation
+    # )
 
-
-@bp.route("/history/clear", methods=["POST"])
-async def clear_messages():
-    ## get the user id from the request headers
-    authenticated_user = get_authenticated_user_details(request_headers=request.headers)
-    user_id = authenticated_user["user_principal_id"]
-
-    ## check request for conversation_id
-    request_json = await request.get_json()
-    conversation_id = request_json.get("conversation_id", None)
-
-    try:
-        if not conversation_id:
-            return jsonify({"error": "conversation_id is required"}), 400
-
-        ## make sure cosmos is configured
-        cosmos_conversation_client = init_cosmosdb_client()
-        if not cosmos_conversation_client:
-            raise Exception("CosmosDB is not configured or not working")
-
-        ## delete the conversation messages from cosmos
-        deleted_messages = await cosmos_conversation_client.delete_messages(
-            conversation_id, user_id
-        )
-
-        return (
-            jsonify(
-                {
-                    "message": "Successfully deleted messages in conversation",
-                    "conversation_id": conversation_id,
-                }
-            ),
-            200,
-        )
-    except Exception as e:
-        logging.exception("Exception in /history/clear_messages")
-        return jsonify({"error": str(e)}), 500
+    # await cosmos_conversation_client.cosmosdb_client.close()
+    # return jsonify(updated_conversation), 200
 
 
-@bp.route("/history/ensure", methods=["GET"])
-async def ensure_cosmos():
-    if not AZURE_COSMOSDB_ACCOUNT:
-        return jsonify({"error": "CosmosDB is not configured"}), 404
+# @bp.route("/history/delete_all", methods=["DELETE"])
+# async def delete_all_conversations():
+#     ## get the user id from the request headers
+#     authenticated_user = get_authenticated_user_details(request_headers=request.headers)
+#     user_id = authenticated_user["user_principal_id"]
 
-    try:
-        cosmos_conversation_client = init_cosmosdb_client()
-        success, err = await cosmos_conversation_client.ensure()
-        if not cosmos_conversation_client or not success:
-            if err:
-                return jsonify({"error": err}), 422
-            return jsonify({"error": "CosmosDB is not configured or not working"}), 500
+#     # get conversations for user
+#     try:
+#         ## make sure cosmos is configured
+#         cosmos_conversation_client = init_cosmosdb_client()
+#         if not cosmos_conversation_client:
+#             raise Exception("CosmosDB is not configured or not working")
 
-        await cosmos_conversation_client.cosmosdb_client.close()
-        return jsonify({"message": "CosmosDB is configured and working"}), 200
-    except Exception as e:
-        logging.exception("Exception in /history/ensure")
-        cosmos_exception = str(e)
-        if "Invalid credentials" in cosmos_exception:
-            return jsonify({"error": cosmos_exception}), 401
-        elif "Invalid CosmosDB database name" in cosmos_exception:
-            return (
-                jsonify(
-                    {
-                        "error": f"{cosmos_exception} {AZURE_COSMOSDB_DATABASE} for account {AZURE_COSMOSDB_ACCOUNT}"
-                    }
-                ),
-                422,
-            )
-        elif "Invalid CosmosDB container name" in cosmos_exception:
-            return (
-                jsonify(
-                    {
-                        "error": f"{cosmos_exception}: {AZURE_COSMOSDB_CONVERSATIONS_CONTAINER}"
-                    }
-                ),
-                422,
-            )
-        else:
-            return jsonify({"error": "CosmosDB is not working"}), 500
+#         conversations = await cosmos_conversation_client.get_conversations(
+#             user_id, offset=0, limit=None
+#         )
+#         if not conversations:
+#             return jsonify({"error": f"No conversations for {user_id} were found"}), 404
+
+#         # delete each conversation
+#         for conversation in conversations:
+#             ## delete the conversation messages from cosmos first
+#             deleted_messages = await cosmos_conversation_client.delete_messages(
+#                 conversation["id"], user_id
+#             )
+
+#             ## Now delete the conversation
+#             deleted_conversation = await cosmos_conversation_client.delete_conversation(
+#                 user_id, conversation["id"]
+#             )
+#         await cosmos_conversation_client.cosmosdb_client.close()
+#         return (
+#             jsonify(
+#                 {
+#                     "message": f"Successfully deleted conversation and messages for user {user_id}"
+#                 }
+#             ),
+#             200,
+#         )
+
+#     except Exception as e:
+#         logging.exception("Exception in /history/delete_all")
+#         return jsonify({"error": str(e)}), 500
 
 
-async def generate_title(conversation_messages):
-    ## make sure the messages are sorted by _ts descending
-    title_prompt = 'Summarize the conversation so far into a 4-word or less title. Do not use any quotation marks or punctuation. Respond with a json object in the format {{"title": string}}. Do not include any other commentary or description.'
+# @bp.route("/history/clear", methods=["POST"])
+# async def clear_messages():
+#     ## get the user id from the request headers
+#     authenticated_user = get_authenticated_user_details(request_headers=request.headers)
+#     user_id = authenticated_user["user_principal_id"]
 
-    messages = [
-        {"role": msg["role"], "content": msg["content"]}
-        for msg in conversation_messages
-    ]
-    messages.append({"role": "user", "content": title_prompt})
+#     ## check request for conversation_id
+#     request_json = await request.get_json()
+#     conversation_id = request_json.get("conversation_id", None)
 
-    try:
-        azure_openai_client = init_openai_client(use_data=False)
-        response = await azure_openai_client.chat.completions.create(
-            model=AZURE_OPENAI_MODEL, messages=messages, temperature=1, max_tokens=64
-        )
+#     try:
+#         if not conversation_id:
+#             return jsonify({"error": "conversation_id is required"}), 400
 
-        title = json.loads(response.choices[0].message.content)["title"]
-        return title
-    except Exception as e:
-        return messages[-2]["content"]
+#         ## make sure cosmos is configured
+#         cosmos_conversation_client = init_cosmosdb_client()
+#         if not cosmos_conversation_client:
+#             raise Exception("CosmosDB is not configured or not working")
+
+#         ## delete the conversation messages from cosmos
+#         deleted_messages = await cosmos_conversation_client.delete_messages(
+#             conversation_id, user_id
+#         )
+
+#         return (
+#             jsonify(
+#                 {
+#                     "message": "Successfully deleted messages in conversation",
+#                     "conversation_id": conversation_id,
+#                 }
+#             ),
+#             200,
+#         )
+#     except Exception as e:
+#         logging.exception("Exception in /history/clear_messages")
+#         return jsonify({"error": str(e)}), 500
+
+
+# @bp.route("/history/ensure", methods=["GET"])
+# async def ensure_cosmos():
+#     if not AZURE_COSMOSDB_ACCOUNT:
+#         return jsonify({"error": "CosmosDB is not configured"}), 404
+
+#     try:
+#         cosmos_conversation_client = init_cosmosdb_client()
+#         success, err = await cosmos_conversation_client.ensure()
+#         if not cosmos_conversation_client or not success:
+#             if err:
+#                 return jsonify({"error": err}), 422
+#             return jsonify({"error": "CosmosDB is not configured or not working"}), 500
+
+#         await cosmos_conversation_client.cosmosdb_client.close()
+#         return jsonify({"message": "CosmosDB is configured and working"}), 200
+#     except Exception as e:
+#         logging.exception("Exception in /history/ensure")
+#         cosmos_exception = str(e)
+#         if "Invalid credentials" in cosmos_exception:
+#             return jsonify({"error": cosmos_exception}), 401
+#         elif "Invalid CosmosDB database name" in cosmos_exception:
+#             return (
+#                 jsonify(
+#                     {
+#                         "error": f"{cosmos_exception} {AZURE_COSMOSDB_DATABASE} for account {AZURE_COSMOSDB_ACCOUNT}"
+#                     }
+#                 ),
+#                 422,
+#             )
+#         elif "Invalid CosmosDB container name" in cosmos_exception:
+#             return (
+#                 jsonify(
+#                     {
+#                         "error": f"{cosmos_exception}: {AZURE_COSMOSDB_CONVERSATIONS_CONTAINER}"
+#                     }
+#                 ),
+#                 422,
+#             )
+#         else:
+#             return jsonify({"error": "CosmosDB is not working"}), 500
+
+
+# async def generate_title(conversation_messages):
+#     ## make sure the messages are sorted by _ts descending
+#     title_prompt = 'Summarize the conversation so far into a 4-word or less title. Do not use any quotation marks or punctuation. Respond with a json object in the format {{"title": string}}. Do not include any other commentary or description.'
+
+#     messages = [
+#         {"role": msg["role"], "content": msg["content"]}
+#         for msg in conversation_messages
+#     ]
+#     messages.append({"role": "user", "content": title_prompt})
+
+#     try:
+#         azure_openai_client = init_openai_client(use_data=False)
+#         response = await azure_openai_client.chat.completions.create(
+#             model=AZURE_OPENAI_MODEL, messages=messages, temperature=1, max_tokens=64
+#         )
+
+#         title = json.loads(response.choices[0].message.content)["title"]
+#         return title
+#     except Exception as e:
+#         return messages[-2]["content"]
 
 app = create_app()
