+++ backend/azure_data/constants.py
@@ -1,0 +1,101 @@
+import os
+from dotenv import load_dotenv
+load_dotenv()
+# Current minimum Azure OpenAI version supported
+MINIMUM_SUPPORTED_AZURE_OPENAI_PREVIEW_API_VERSION = "2024-02-15-preview"
+USER_AGENT = "GitHubSampleWebApp/AsyncAzureOpenAI/1.0.0"
+
+# On Your Data Settings
+DATASOURCE_TYPE = os.environ.get("DATASOURCE_TYPE", "AzureCognitiveSearch")
+SEARCH_TOP_K = os.environ.get("SEARCH_TOP_K", 5)
+SEARCH_STRICTNESS = os.environ.get("SEARCH_STRICTNESS", 3)
+SEARCH_ENABLE_IN_DOMAIN = os.environ.get("SEARCH_ENABLE_IN_DOMAIN", "true")
+
+# ACS Integration Settings
+AZURE_SEARCH_SERVICE = os.environ.get("AZURE_SEARCH_SERVICE")
+AZURE_SEARCH_INDEX = os.environ.get("AZURE_SEARCH_INDEX")
+AZURE_SEARCH_KEY = os.environ.get("AZURE_SEARCH_KEY", None)
+AZURE_SEARCH_USE_SEMANTIC_SEARCH = os.environ.get(
+    "AZURE_SEARCH_USE_SEMANTIC_SEARCH", "false"
+)
+AZURE_SEARCH_SEMANTIC_SEARCH_CONFIG = os.environ.get(
+    "AZURE_SEARCH_SEMANTIC_SEARCH_CONFIG", "default"
+)
+AZURE_SEARCH_TOP_K = os.environ.get("AZURE_SEARCH_TOP_K", SEARCH_TOP_K)
+AZURE_SEARCH_ENABLE_IN_DOMAIN = os.environ.get(
+    "AZURE_SEARCH_ENABLE_IN_DOMAIN", SEARCH_ENABLE_IN_DOMAIN
+)
+AZURE_SEARCH_CONTENT_COLUMNS = os.environ.get("AZURE_SEARCH_CONTENT_COLUMNS")
+AZURE_SEARCH_FILENAME_COLUMN = os.environ.get("AZURE_SEARCH_FILENAME_COLUMN")
+AZURE_SEARCH_TITLE_COLUMN = os.environ.get("AZURE_SEARCH_TITLE_COLUMN")
+AZURE_SEARCH_URL_COLUMN = os.environ.get("AZURE_SEARCH_URL_COLUMN")
+AZURE_SEARCH_VECTOR_COLUMNS = os.environ.get("AZURE_SEARCH_VECTOR_COLUMNS")
+AZURE_SEARCH_QUERY_TYPE = os.environ.get("AZURE_SEARCH_QUERY_TYPE")
+AZURE_SEARCH_PERMITTED_GROUPS_COLUMN = os.environ.get(
+    "AZURE_SEARCH_PERMITTED_GROUPS_COLUMN"
+)
+AZURE_SEARCH_STRICTNESS = os.environ.get("AZURE_SEARCH_STRICTNESS", SEARCH_STRICTNESS)
+
+# AOAI Integration Settings
+AZURE_OPENAI_RESOURCE = os.environ.get("AZURE_OPENAI_RESOURCE")
+AZURE_OPENAI_MODEL = os.environ.get("AZURE_OPENAI_MODEL")
+AZURE_OPENAI_ENDPOINT = os.environ.get("AZURE_OPENAI_ENDPOINT")
+AZURE_OPENAI_KEY = os.environ.get("AZURE_OPENAI_KEY")
+AZURE_OPENAI_TEMPERATURE = os.environ.get("AZURE_OPENAI_TEMPERATURE", 0)
+AZURE_OPENAI_TOP_P = os.environ.get("AZURE_OPENAI_TOP_P", 1.0)
+AZURE_OPENAI_MAX_TOKENS = os.environ.get("AZURE_OPENAI_MAX_TOKENS", 1000)
+AZURE_OPENAI_STOP_SEQUENCE = os.environ.get("AZURE_OPENAI_STOP_SEQUENCE")
+# AZURE_OPENAI_SYSTEM_MESSAGE = os.environ.get(
+#     "AZURE_OPENAI_SYSTEM_MESSAGE",
+#     "You are an AI assistant that helps people find information.",
+# )
+SYSTEM_MESSAGE= os.environ.get(
+    "SYSTEM_MESSAGE",
+    "You are an AI assistant that helps people find information on software development languages like .NET, Java, C, C-sharp, JavaScript, Python\
+    When providing an answer, make sure to include all the necessary technical details and context.",
+)
+AZURE_OPENAI_PREVIEW_API_VERSION = os.environ.get(
+    "AZURE_OPENAI_PREVIEW_API_VERSION",
+    MINIMUM_SUPPORTED_AZURE_OPENAI_PREVIEW_API_VERSION,
+)
+AZURE_OPENAI_STREAM = os.environ.get("AZURE_OPENAI_STREAM", "true")
+AZURE_OPENAI_MODEL_NAME = os.environ.get(
+    "AZURE_OPENAI_MODEL_NAME", "gpt-35-turbo-16k"
+)  # Name of the model, e.g. 'gpt-35-turbo-16k' or 'gpt-4'
+AZURE_OPENAI_EMBEDDING_ENDPOINT = os.environ.get("AZURE_OPENAI_EMBEDDING_ENDPOINT")
+AZURE_OPENAI_EMBEDDING_KEY = os.environ.get("AZURE_OPENAI_EMBEDDING_KEY")
+AZURE_OPENAI_EMBEDDING_NAME = os.environ.get("AZURE_OPENAI_EMBEDDING_NAME", "")
+
+SHOULD_STREAM = True if AZURE_OPENAI_STREAM.lower() == "true" else False
+
+# Chat History CosmosDB Integration Settings
+AZURE_COSMOSDB_DATABASE = os.environ.get("AZURE_COSMOSDB_DATABASE")
+AZURE_COSMOSDB_ACCOUNT = os.environ.get("AZURE_COSMOSDB_ACCOUNT")
+AZURE_COSMOSDB_CONVERSATIONS_CONTAINER = os.environ.get(
+    "AZURE_COSMOSDB_CONVERSATIONS_CONTAINER"
+)
+AZURE_COSMOSDB_ACCOUNT_KEY = os.environ.get("AZURE_COSMOSDB_ACCOUNT_KEY")
+AZURE_COSMOSDB_ENABLE_FEEDBACK = (
+    os.environ.get("AZURE_COSMOSDB_ENABLE_FEEDBACK", "false").lower() == "true"
+)
+
+# Promptflow Integration Settings
+USE_PROMPTFLOW = os.environ.get("USE_PROMPTFLOW", "false").lower() == "true"
+PROMPTFLOW_ENDPOINT = os.environ.get("PROMPTFLOW_ENDPOINT")
+PROMPTFLOW_API_KEY = os.environ.get("PROMPTFLOW_API_KEY")
+PROMPTFLOW_RESPONSE_TIMEOUT = os.environ.get("PROMPTFLOW_RESPONSE_TIMEOUT", 30.0)
+# default request and response field names are input -> 'query' and output -> 'reply'
+PROMPTFLOW_REQUEST_FIELD_NAME = os.environ.get("PROMPTFLOW_REQUEST_FIELD_NAME", "query")
+PROMPTFLOW_RESPONSE_FIELD_NAME = os.environ.get(
+    "PROMPTFLOW_RESPONSE_FIELD_NAME", "reply"
+)
+PROMPTFLOW_CITATIONS_FIELD_NAME = os.environ.get(
+    "PROMPTFLOW_CITATIONS_FIELD_NAME", "documents"
+)
+# Frontend Settings via Environment Variables
+AUTH_ENABLED = os.environ.get("AUTH_ENABLED", "true").lower() == "true"
+CHAT_HISTORY_ENABLED = (
+    AZURE_COSMOSDB_ACCOUNT
+    and AZURE_COSMOSDB_DATABASE
+    and AZURE_COSMOSDB_CONVERSATIONS_CONTAINER
+)
