+++ requirements-dev.txt
@@ -1,13 +1,13 @@
 -r requirements.txt
-azure-ai-formrecognizer==3.2.1
-Markdown==3.4.4
-requests==2.31.0
-tqdm==4.66.1
-tiktoken==0.4.0
-langchain==0.0.340
-bs4==0.0.1
-urllib3==2.1.0
-pytest==7.4.0
-pytest-asyncio==0.23.2
+azure-ai-formrecognizer
+Markdown
+requests
+tqdm
+tiktoken
+langchain
+bs4
+urllib3
+pytest
+pytest-asyncio
 azure-storage-blob
 chardet
